# openapi.yaml
openapi: 3.0.1
info:
  title: CG Assignment
  description: >-
    This is an assignment for Account Balance. For this sample, you can use the
    api key `special-key` to test the authorization     filters.
  termsOfService: 'http://swagger.io/terms/'
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.0
servers:
  - url: 'https://localhost:8080/v1'
  - url: 'http://localhost:8080/v1'
paths:
  /accounts:
    get:
      summary: Get the list of accounts with filtering feature
      operationId: accountList
      parameters:
        - name: accountType
          in: query
          description: Account type values that need to be consider for filters
          required: false
          schema:
            type: string
        - name: crrCode
          in: query
          description: Currenncy code values that need to be consider for filters
          required: false
          schema:
            type: string
        - name: countCode
          in: query
          description: Country code values that need to be consider for filters
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Account'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No record found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        security:
          - api_key: []
  '/accounts/{accountNumber}':
    get:
      summary: Get account detail by Account Number
      operationId: accountByAccountNumber
      parameters:
        - name: accountNumber
          in: path
          description: Enter valid account number
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No record found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        security:
          - api_key: []
components:
  schemas:
    Customer:
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
    Account:
      type: object
      properties:
        id:
          type: integer
          format: int64
        accountNumber:
          type: string
        accountType:
          type: string
          description: Different account type i.e - CHECKING / SAVING
        accountHolderName:
          type: string
        bankName:
          type: string
        country:
          type: string
        currency:
          type: string
        customer:
          $ref: '#/components/schemas/Customer'
        bankCode:
          type: string
        branchCode:
          type: string
        status:
          type: string
          description: Define status of the account NEW / VERIFIED / FAILED / DORMENT
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        details:
          type: string
  securitySchemes:
    accounts_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: ''
          scopes:
            'write:accounts': modify accounts
            'read:accounts': read your accounts
    api_key:
      type: apiKey
      name: api_key
      in: header