# spring-server-generated/src/main/java/io/swagger/api/AccountsApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.24).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.swagger.api;

import java.util.List;

import javax.validation.Valid;

import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;

import io.swagger.model.Account;
import io.swagger.model.ErrorResponse;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.responses.ApiResponses;

@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2021-02-23T04:43:46.613Z[GMT]")
public interface AccountsApi {

	@Operation(summary = "Get account detail by Account Number", description = "", tags = {})
	@ApiResponses(value = {
			@ApiResponse(responseCode = "200", description = "Successful operation", content = @Content(schema = @Schema(implementation = Account.class))),

			@ApiResponse(responseCode = "400", description = "Bad request", content = @Content(schema = @Schema(implementation = ErrorResponse.class))),

			@ApiResponse(responseCode = "404", description = "No record found", content = @Content(schema = @Schema(implementation = ErrorResponse.class))) })
	@RequestMapping(value = "/accounts/{accountNumber}", produces = { "application/json" }, method = RequestMethod.GET)
	ResponseEntity<Account> accountByAccountNumber(
			@Parameter(in = ParameterIn.PATH, description = "Enter valid account number", required = true, schema = @Schema()) @PathVariable("accountNumber") String accountNumber);

	@Operation(summary = "Get the list of accounts with filtering feature", description = "", tags = {})
	@ApiResponses(value = {
			@ApiResponse(responseCode = "200", description = "Successful operation", content = @Content(array = @ArraySchema(schema = @Schema(implementation = Account.class)))),

			@ApiResponse(responseCode = "400", description = "Bad request", content = @Content(schema = @Schema(implementation = ErrorResponse.class))),

			@ApiResponse(responseCode = "404", description = "No record found", content = @Content(schema = @Schema(implementation = ErrorResponse.class))) })
	@RequestMapping(value = "/accounts", produces = { "application/json" }, method = RequestMethod.GET)
	ResponseEntity<List<Account>> accountList(
			@Parameter(in = ParameterIn.QUERY, description = "Account type values that need to be consider for filters", schema = @Schema()) @Valid @RequestParam(value = "accountType", required = false) String accountType,
			@Parameter(in = ParameterIn.QUERY, description = "Currenncy code values that need to be consider for filters", schema = @Schema()) @Valid @RequestParam(value = "crrCode", required = false) String crrCode,
			@Parameter(in = ParameterIn.QUERY, description = "Country code values that need to be consider for filters", schema = @Schema()) @Valid @RequestParam(value = "countCode", required = false) String countCode);

}